requirements.txt
Flask
Flask-Cors
Flask-Mail
python-dotenv